+++ packages/boxes/minimum-fee/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rosen-bridge/minimum-fee
 
+## 2.0.0
+
+### Major Changes
+
+- Update api calls to use token and remove required address
+
+### Patch Changes
+
+- update ergo explorer and node clients
+
 ## 1.0.2
 
 ### Patch Changes