+++ packages/boxes/rwt-repo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rosen-bridge/rwt-repo
 
+## 2.0.2
+
+### Patch Changes
+
+- update ergo explorer and node clients
+
 ## 2.0.1
 
 ### Patch Changes